Exception handled too slow StateMachine

+++ TUeLDT/include/LaneFilter.h
@@ -55,10 +55,10 @@ public:	 	//Public Interface
   		const int 	PURVIEW_LINE_ICCS;	/**< Base line in Image-Center-CS [pixel-lines] */
 
 		const float	BINS_STEP_cm;  		/**< Step size for the Histogram BINS [cm] */
-  		const size_t	BINS_MAX_cm;		/**< Max value of the #BINS_cm **/
+  		const float	BINS_MAX_cm;		/**< Max value of the #BINS_cm **/
 		const size_t   	COUNT_BINS;   		/**< Number of bins in each Histogram */
 
-		const VectorXi	BINS_cm;		/**<Describes the Histogram BINS in the Vehicle-Symmetry-CS [cm] */
+		const VectorXf	BINS_cm;		/**<Describes the Histogram BINS in the Vehicle-Symmetry-CS [cm] */
 
 		
 		const float	BASE_STEP;		/**< Step size for the base bins */
